Duplicate Oomox exporter with backgrounds flipped
A lot of programs end up using the textview background color instead of
the actual gtk background for displays, so added an alternate 'flipped'
background where the base is color8 and everythign else is color0, as
for many programs this is more desireable since the theme is built with
color0/color7 combo in mind. Made a separate exporter since honestly
it's a completely personal preference.